pwaudiosink: Add use-global-probed-caps-cache property
If the PipeWire graph never changes, or at least changes only in ways
that don't affect what formats can be played, then it makes sense to
cache probed caps not just locally, but globally, since then, future
pwaudiosink instances don't have to redo the probing unnecessarily.

This is only okay to do if the probing results remain the same all the
time. If they can change, for example, because sink nodes get added or
removed on the fly to/from the PipeWire graph, then such a global cache
can lead to incorrect behavior. For this reason, the property's
default value is set to FALSE.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>