pwaudiosink: Add stream-activated GLib GObject signal
This signal is useful for cases where upstream needs some sort of extra
synchronization based on when the pw_stream actually begins to invoke
the process callback. In particular, when the pipeline clock is not the
pwaudiosink stream clock, this can be important to do for avoiding losing
the first N milliseconds worth of audio.

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>