pwaudioformat: Add support for AC3, EAC3, TrueHD, DTS, MPEG-H
This adds support for these compressed formats by detecting their headers.
It is done that way to make it straightforward to support more compressed
formats in the future, should more be added in PipeWire.

For the same reason, NO_CHANNELS macros are added instead of making
special cases for MPEG-H (due to its lack of channel count); going
with NO_CHANNELS macros is more generic, and useful if in the future,
some other format that has no channel count is added to PipeWire.

MPEG-H support was introduced in GStreamer 1.28. The caps used here
follow GStreamer 1.28's design. For more about this, look at:

https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/9361#note_3174408

Signed-off-by: Carlos Rafael Giani <crg7475@mailbox.org>